feat(bills): render bills at the app root; drop camera frosting
Phase B of moving bills out of the scanner. A presented bill now renders at the
app root instead of inside the camera screen, so it appears over ANY screen and
is fully decoupled from the camera.

- New BillOverlay (:shared:bills), mounted at the app root in App.kt next to
  ScrimOverlay. It reads the app-scoped billState via LocalSessionController and
  owns the swipe-to-dismiss gesture, enter/exit animation, tip-modal positioning,
  and the below-bill decorators. Touch-transparent when no bill is present.
- ScannableContainer slims down to camera + HUD: it no longer draws the bill,
  and hides its DecorView HUD while a bill is up (billState.bill != null).
- Remove the frosted tip-card camera backdrop entirely: delete
  TipCardCameraBackdrop, drop the previewView plumbing from Scanner, and render
  TipCard as a solid opaque card (LocalTipCard{Color,BaseAlpha} now default to
  the opaque fallback; the backdrop composition local is gone).

The FrostedTipCard feature flag is now unused (left in place; harmless).
