+++ apps/flipcash/app/build.gradle.kts
@@ -183,6 +183,7 @@ dependencies {
 
     implementation(project(":apps:flipcash:core"))
     implementation(project(":apps:flipcash:shared:accesskey"))
+    implementation(project(":apps:flipcash:shared:bills"))
     implementation(project(":apps:flipcash:shared:analytics"))
     implementation(project(":apps:flipcash:shared:appsettings"))
     implementation(project(":apps:flipcash:shared:appupdates"))
